+++ app/src/main/java/org/oxycblt/auxio/widgets/Forms.kt
@@ -30,54 +30,47 @@ import org.oxycblt.auxio.util.newMainIntent
  * The default widget is displayed whenever there is no music playing. It just shows the message "No
  * music playing".
  */
-fun createDefaultWidget(context: Context): RemoteViews {
-    return createViews(context, R.layout.widget_default)
-}
+fun createDefaultWidget(context: Context) = createViews(context, R.layout.widget_default)
 
 /**
  * The tiny widget is for an edge-case situation where a widget falls under the size class of the
  * small widget, either via landscape mode or exceptionally small screens.
  */
-fun createTinyWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews {
-    return createViews(context, R.layout.widget_tiny)
+fun createTinyWidget(context: Context, state: WidgetComponent.WidgetState) =
+    createViews(context, R.layout.widget_tiny)
         .applyMeta(context, state)
         .applyBasicControls(context, state)
-}
 
 /**
  * The small widget is for 2x2 widgets and just shows the cover art and playback controls. This is
  * generally because a Medium widget is too large for this widget size and a text-only widget is too
  * small for this widget size.
  */
-fun createSmallWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews {
-    return createViews(context, R.layout.widget_small)
+fun createSmallWidget(context: Context, state: WidgetComponent.WidgetState) =
+    createViews(context, R.layout.widget_small)
         .applyCover(context, state)
         .applyBasicControls(context, state)
-}
 
 /**
  * The medium widget is for 2x3 widgets and shows the cover art, title/artist, and three controls.
  * This is the default widget configuration.
  */
-fun createMediumWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews {
-    return createViews(context, R.layout.widget_medium)
+fun createMediumWidget(context: Context, state: WidgetComponent.WidgetState) =
+    createViews(context, R.layout.widget_medium)
         .applyMeta(context, state)
         .applyBasicControls(context, state)
-}
 
 /** The wide widget is for Nx2 widgets and is like the small widget but with more controls. */
-fun createWideWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews {
-    return createViews(context, R.layout.widget_wide)
+fun createWideWidget(context: Context, state: WidgetComponent.WidgetState) =
+    createViews(context, R.layout.widget_wide)
         .applyCover(context, state)
         .applyFullControls(context, state)
-}
 
 /** The large widget is for 3x4 widgets and shows all metadata and controls. */
-fun createLargeWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews {
-    return createViews(context, R.layout.widget_large)
+fun createLargeWidget(context: Context, state: WidgetComponent.WidgetState): RemoteViews =
+    createViews(context, R.layout.widget_large)
         .applyMeta(context, state)
         .applyFullControls(context, state)
-}
 
 private fun createViews(context: Context, @LayoutRes layout: Int): RemoteViews {
     val views = RemoteViews(context.packageName, layout)
@@ -114,7 +107,7 @@ private fun RemoteViews.applyCover(
     return this
 }
 
-private fun RemoteViews.applyPlayControls(
+private fun RemoteViews.applyBasicControls(
     context: Context,
     state: WidgetComponent.WidgetState
 ): RemoteViews {
@@ -128,15 +121,6 @@ private fun RemoteViews.applyPlayControls(
         } else {
             R.drawable.ic_play
         })
-
-    return this
-}
-
-private fun RemoteViews.applyBasicControls(
-    context: Context,
-    state: WidgetComponent.WidgetState
-): RemoteViews {
-    applyPlayControls(context, state)
 
     setOnClickPendingIntent(
         R.id.widget_skip_prev, context.newBroadcastIntent(PlaybackService.ACTION_SKIP_PREV))